ft_calloc.c fixed (not filling last byte of array), and fixed some functions (using int instead of size_t)

+++ ft_atoi.c
@@ -6,15 +6,15 @@
 /*   By: ncofre <ncofre@student.42.us.org>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/07 00:01:51 by ncofre            #+#    #+#             */
-/*   Updated: 2020/11/07 08:32:33 by ncofre           ###   ########.fr       */
+/*   Updated: 2020/11/22 22:00:44 by ncofre           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 int	ft_atoi(const char *nptr)
 {
-	int i;
+	long i;
 	int sign;
-	int number;
+	long number;
 
 	i = 0;
 	sign = 1;
